:bug: Fix user credentials display in main component and add new error handler on backend

+++ backend_spring/src/main/java/com/example/fcmanager/feature/user/service/UserServiceDefault.java
@@ -1,14 +1,15 @@
 package com.example.fcmanager.feature.user.service;
 
+import java.util.UUID;
+
+import org.springframework.stereotype.Service;
+
+import com.example.fcmanager.exception.UserNotFoundWithIdException;
 import com.example.fcmanager.feature.user.dto.UserResponseDto;
 import com.example.fcmanager.feature.user.mapper.UserMapper;
 import com.example.fcmanager.feature.user.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 
-import java.util.List;
-import java.util.UUID;
-import java.util.stream.Collectors;
+import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
@@ -17,21 +18,17 @@ public class UserServiceDefault implements UserService {
     private final UserMapper userMapper;
 
     @Override
-    public List<UserResponseDto> getAllUsers() {
-        return userRepository.findAll().stream()
-                .map(userMapper::toUserDto)
-                .collect(Collectors.toList());
-    }
-
-    @Override
     public UserResponseDto getUserById(UUID id) {
         return userRepository.findById(id)
                 .map(userMapper::toUserDto)
-                .orElseThrow(() -> new RuntimeException("User not found"));
+                .orElseThrow(() -> new UserNotFoundWithIdException("User not found with id: " + id));
     }
 
     @Override
     public void deleteUser(UUID id) {
+        if (!userRepository.existsById(id)) {
+            throw new UserNotFoundWithIdException("Cannot delete. User not found with id: " + id);
+        }
         userRepository.deleteById(id);
     }
 }
